exception handling in delete method

+++ src/main/java/com/philimonov/spring/rest/controller/MyRESTController.java
@@ -51,6 +51,10 @@ public class MyRESTController {
 
     @DeleteMapping("/employees/{id}")
     public String deleteEmployee(@PathVariable int id) {
+        Employee employee = employeeService.getEmployee(id);
+        if(employee == null){
+            throw new NoSuchEmployeeException("There is no employee with ID = " + id + " in DataBase");
+        }
         employeeService.deleteEmployee(id);
         return "Employee with id = " + id + " was deleted!!!";
     }
